Refactors and streamlines image handling
Removes task scheduling and related worker functionalities.

Consolidates file conversion utilities and image handling, simplifying codebase and improving maintainability.

Moves access control logic directly into character retrieval to ensure proper authorization.

Updates session configuration for improved security and consistency.
