+++ .github/workflows/build-and-release.yml
@@ -49,7 +49,7 @@ jobs:
               uses: docker/build-push-action@v5.3.0
               with:
                   context: .
-                  platforms: linux/amd64,linux/arm64
+                  platforms: linux/amd64
                   file: Dockerfile
                   push: true
                   tags: ${{ steps.metadata.outputs.tags }}
