+++ nuxt.config.ts
@@ -73,11 +73,13 @@ export default defineNuxtConfig({
 		// Caching
 		ttl: 1000 * 60 * 60 * 24, // In milliseconds, default 24 Hours
 		redisURL: '', // optional, redis://<REDIS_USER>:<REDIS_PASSWORD>@<REDIS_HOST>:<REDIS_PORT>
+		redisTLS: true,
 		// Session configuration for nuxt-auth-utils
 		session: {
+			password: process.env.SESSION_PASSWORD ?? '',
 			maxAge: 60 * 60 * 24 * 7, // 1 week
-			cookieName: 'auth_session',
-			cookieAttributes: {
+			name: 'auth_session',
+			cookie: {
 				secure: process.env.NODE_ENV === 'production',
 				sameSite: 'lax',
 				httpOnly: true,
