+++ server/utils/characterCardUtilities.ts
