+++ server/utils/imageHandling.ts
@@ -10,7 +10,7 @@ export async function loadImageById(id: number): Promise<Uint8Array> {
 			return value;
 		}
 
-		const file = await fs.readFile(path.join(runtimeConfig.imageFolder, `${id}.png`));
+		const file = await fs.readFile(path.join('./images', `${id}.png`));
 		await useCache().setItemRaw(id.toString(), file);
 
 		return file;
@@ -21,7 +21,7 @@ export async function loadImageById(id: number): Promise<Uint8Array> {
 
 export async function saveImageById(id: number, data: Uint8Array): Promise<void> {
 	try {
-		await fs.writeFile(path.join(runtimeConfig.imageFolder, `${id}.png`), data);
+		await fs.writeFile(path.join('./images', `${id}.png`), data);
 		await useCache().setItemRaw(id.toString(), data);
 	} catch (error: any) {
 		throw new Error(`Failed to save image to file system: ${error.message}`);
@@ -30,7 +30,7 @@ export async function saveImageById(id: number, data: Uint8Array): Promise<void>
 
 export async function deleteImageById(id: number): Promise<void> {
 	try {
-		await fs.unlink(path.join(runtimeConfig.imageFolder, `${id}.png`));
+		await fs.unlink(path.join('./images', `${id}.png`));
 	} catch (error: any) {
 		throw new Error(`Failed to save image to file system: ${error.message}`);
 	}
